Move RestoreClient from cmd to pkg package
